gait_project/requirements.txt
streamlit
torch
torchvision
pillow
matplotlib
pandas
scikit-learn
scipy